refactor(graph-gateway): relocate indexers pois merge_queries logic
